Refactor BSON usage in gRPC server methods
This commit updates the gRPC server implementation for the BattleService by replacing instances of `primitive.M` with `bson.M` for better consistency and clarity. Additionally, it modifies the way timestamps are set in the `UpdateParticipantStats` method to use `time.Now()` directly, enhancing the accuracy of update timestamps. These changes improve code readability and maintainability.